refactor: move category router to new module structure and update imports
